Fix unbounded send-epoch catch-up loop and I2C HAL type mismatch
Bug #5: Replace unbounded while loop in g_next_send_epoch catch-up
(main.c:1171) with O(1) arithmetic. A corrupted/stale BKP value
could cause millions of iterations, starving the watchdog.

Bug #1/#2: Widen sensirion_i2c_hal_read/write count parameter from
uint8_t to uint16_t to match callers in sensirion_i2c.c which pass
uint16_t size values. Prevents silent truncation on larger reads.

+++ Core/Src/sensirion/sensirion_i2c_hal.c
@@ -16,14 +16,14 @@ void sensirion_i2c_hal_free(void) {
     /* nothing to free */
 }
 
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
     watchdog_kick();
     HAL_StatusTypeDef status = HAL_I2C_Master_Receive(&hi2c1, (uint16_t)(address << 1), data, count, 1000);
     watchdog_kick();
     return (status == HAL_OK) ? 0 : -1;
 }
 
-int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data, uint8_t count) {
+int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data, uint16_t count) {
     watchdog_kick();
     HAL_StatusTypeDef status = HAL_I2C_Master_Transmit(&hi2c1, (uint16_t)(address << 1), (uint8_t*)data, count, 1000);
     watchdog_kick();
